Supply chain Management
We are Optimizing Supply Chain Management using ML to Address Supply Chain Disruptions in Instant Noodles Delivery

Problem Description:

An FMCG company entered into the instant noodles business two years back. Their higher management has noticed that there is a mismatch in the demand and supply. Where the demand is high, supply is pretty low and where the demand is low, supply is pretty high. In both ways, it is an inventory cost loss to the company; hence, the higher management wants to optimize the supply quantity in every warehouse in the entire country.

Goal and Objective:

Need to build a model, using historical data that will determine the optimum weight of the product to be shipped each time to the warehouse.

Also, need to analyze the demand pattern in different pockets of the country so that management can drive the advertisement campaign, particularly in those pockets.

As a first phase of agreement, the company has shared very limited information. Once we showcase a tangible impact with this much information then the company will open the 360-degree data lake for us to build a more robust model.